refactor(python): Set up `utils` module
