+++ py-polars/docs/source/reference/functions.rst
@@ -42,7 +42,6 @@ Parallelization
    :toctree: api/
 
    collect_all
-   threadpool_size
 
 StringCache
 ~~~~~~~~~~~
