+++ py-polars/docs/source/reference/utils.rst
@@ -8,4 +8,6 @@ Utils
    :toctree: api/
 
     build_info
+    get_index_type
     show_versions
+    threadpool_size
