+++ py-polars/polars/convert.py
@@ -8,7 +8,7 @@ from polars.dependencies import pandas as pd
 from polars.dependencies import pyarrow as pa
 from polars.exceptions import NoDataError
 from polars.internals import DataFrame, Series
-from polars.utils import deprecate_nonkeyword_arguments, deprecated_alias
+from polars.utils.decorators import deprecate_nonkeyword_arguments, deprecated_alias
 
 if TYPE_CHECKING:
     from polars.datatypes import SchemaDefinition, SchemaDict
