+++ py-polars/polars/internals/dataframe/groupby.py
@@ -10,7 +10,8 @@ from typing import (
 )
 
 import polars.internals as pli
-from polars.utils import _timedelta_to_pl_duration, deprecated_alias, redirect
+from polars.utils.convert import _timedelta_to_pl_duration
+from polars.utils.decorators import deprecated_alias, redirect
 
 if TYPE_CHECKING:
     from datetime import timedelta
