+++ py-polars/polars/internals/series/list.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, Callable
 
 import polars.internals as pli
 from polars.internals.series.utils import expr_dispatch
-from polars.utils import deprecate_nonkeyword_arguments
+from polars.utils.decorators import deprecate_nonkeyword_arguments
 
 if TYPE_CHECKING:
     from datetime import date, datetime, time
