+++ py-polars/polars/io/avro.py
@@ -4,7 +4,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING, BinaryIO
 
 from polars.internals import DataFrame
-from polars.utils import deprecate_nonkeyword_arguments, normalise_filepath
+from polars.utils.decorators import deprecate_nonkeyword_arguments
+from polars.utils.various import normalise_filepath
 
 if TYPE_CHECKING:
     from io import BytesIO
