+++ py-polars/polars/io/ipc.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Any, BinaryIO
 from polars.dependencies import _PYARROW_AVAILABLE
 from polars.internals import DataFrame, LazyFrame
 from polars.internals.io import _prepare_file_arg
-from polars.utils import deprecate_nonkeyword_arguments
+from polars.utils.decorators import deprecate_nonkeyword_arguments
 
 if TYPE_CHECKING:
     from io import BytesIO
