+++ py-polars/polars/utils/__init__.py
@@ -1,0 +1,33 @@
+"""
+Utility functions.
+
+Functions that are part of the public API are re-exported here.
+"""
+from polars.utils.build_info import build_info
+from polars.utils.convert import (
+    _date_to_pl_date,
+    _time_to_pl_time,
+    _timedelta_to_pl_timedelta,
+    _to_python_datetime,
+    _to_python_decimal,
+    _to_python_time,
+    _to_python_timedelta,
+)
+from polars.utils.meta import get_idx_type, get_index_type, threadpool_size
+from polars.utils.show_versions import show_versions
+
+__all__ = [
+    "build_info",
+    "show_versions",
+    "get_idx_type",
+    "get_index_type",
+    "threadpool_size",
+    # Required for Rust bindings
+    "_date_to_pl_date",
+    "_time_to_pl_time",
+    "_timedelta_to_pl_timedelta",
+    "_to_python_datetime",
+    "_to_python_decimal",
+    "_to_python_time",
+    "_to_python_timedelta",
+]
