+++ py-polars/polars/utils/build_info.py
@@ -2,26 +2,21 @@ from __future__ import annotations
 
 from typing import Any
 
-try:
-    from polars.polars import version
-
-    _version_ = version()
-except ImportError:
-    _version_ = "<missing>"
+from polars.utils.polars_version import get_polars_version
 
 try:
     from polars.polars import _build_info_
 except ImportError:
     _build_info_ = {}
 
-_build_info_["version"] = _version_
+_build_info_["version"] = get_polars_version() or "<missing>"
 
 
 def build_info() -> dict[str, Any]:
     """
-    Return a dict with polars build information.
+    Return a dict with Polars build information.
 
-    If polars was compiled with "build_info" feature gate return the full build info,
+    If Polars was compiled with "build_info" feature gate return the full build info,
     otherwise only version is included. The full build information dict contains
     the following keys ['build', 'info-time', 'dependencies', 'features', 'host',
     'target', 'git', 'version'].
